refactor(BA-6670): move deployment strategy schemas down to common/schema
Move the deployment strategy schemas (RollingUpdateSpec, BlueGreenSpec,
CanarySpec, DeploymentStrategySchema) together with their value types
(ReplicaGroupRolloutSpec, TargetGroupSpec, TrafficStep, TrafficStepInput)
and IntOrPercent into a new shared common/schema module. The manager data
layer no longer imports manager models, making data a pure leaf and cutting
the models<->data import cycle.
