+++ changes/12630.enhance.md
@@ -1,0 +1,1 @@
+Move deployment strategy schemas and their value types from the manager models/data layers down to a new shared `common/schema` module to cut the models↔data import cycle.
