+++ src/ai/backend/common/schema/CLAUDE.md
@@ -1,0 +1,1 @@
+AGENTS.md
