+++ src/ai/backend/manager/data/deployment/modifier.py
@@ -4,7 +4,7 @@ from dataclasses import dataclass, field
 from typing import Any, override
 
 from ai.backend.common.data.model_deployment.types import DeploymentStrategy
-from ai.backend.manager.models.deployment_policy import BlueGreenSpec, RollingUpdateSpec
+from ai.backend.common.schema.deployment import BlueGreenSpec, RollingUpdateSpec
 from ai.backend.manager.types import OptionalState, PartialModifier
