+++ src/ai/backend/manager/models/deployment_policy/__init__.py
@@ -1,15 +1,3 @@
 from .row import DeploymentPolicyRow
-from .schema import (
-    BlueGreenSpec,
-    CanarySpec,
-    DeploymentStrategySchema,
-    RollingUpdateSpec,
-)
 
-__all__ = (
-    "BlueGreenSpec",
-    "CanarySpec",
-    "DeploymentPolicyRow",
-    "DeploymentStrategySchema",
-    "RollingUpdateSpec",
-)
+__all__ = ("DeploymentPolicyRow",)
