+++ src/ai/backend/manager/models/replica_group/row.py
@@ -10,10 +10,10 @@ from sqlalchemy.orm import Mapped, foreign, mapped_column, relationship
 from ai.backend.common.identifier.deployment import DeploymentID
 from ai.backend.common.identifier.deployment_revision import DeploymentRevisionID
 from ai.backend.common.identifier.replica_group import ReplicaGroupID
+from ai.backend.common.schema.deployment import ReplicaGroupRolloutSpec
 from ai.backend.logging import BraceStyleAdapter
 from ai.backend.manager.data.deployment.types import (
     ReplicaGroupLifecycle,
-    ReplicaGroupRolloutSpec,
     ReplicaGroupScalingStatus,
 )
 from ai.backend.manager.models.base import GUID, Base, PydanticColumn, StrEnumType
