+++ tests/unit/manager/models/test_routing_row.py
@@ -9,6 +9,7 @@ from ai.backend.manager.data.deployment.types import (
 )
 from ai.backend.manager.models.agent import AgentRow
 from ai.backend.manager.models.endpoint import EndpointRow
+from ai.backend.manager.models.image import ImageRow
 
 # ORM cluster registration: configure_mappers() (triggered when this isolated
 # test registers a domain-cluster row) resolves string relationships against the
@@ -20,6 +21,7 @@ from ai.backend.manager.models.scaling_group import ScalingGroupForProjectRow
 _ORM_CLUSTER = (
     AgentRow,
     EndpointRow,
+    ImageRow,
     ScalingGroupForProjectRow,
 )
